+++ ui/src/main/java/com/kotcrab/vis/ui/widget/PopupMenu.java
@@ -149,6 +149,18 @@ public class PopupMenu extends Table {
 					}
 				}
 			}
+
+			@Override
+			public void exit (InputEvent event, float x, float y, int pointer, Actor toActor) {
+				if (pointer == -1 && event.getListenerActor() instanceof MenuItem) {
+					if (getRootMenu().subMenuStructureContains(event.getStageX(), event.getStageY())) return;
+
+					MenuItem item = (MenuItem) event.getListenerActor();
+					if (item == activeItem) {
+						setActiveItem(null, false);
+					}
+				}
+			}
 		};
 
 		sharedMenuItemChangeListener = new ChangeListener() {